RosBridge: Move SingletonMeta out of helper.py
Avoids circular imports
